#3 added login error alerts

+++ api/api_set.php
@@ -6,11 +6,6 @@
         switch ($_GET['q']) {
 
             case 'reservations':
-                if (!isset($_SESSION['account']) || !has_permission('mostra_prenotazioni', $_SESSION['account']['cod_gruppo'])) {
-                    header('Location: errors/forbidden.php');
-                    exit();
-                }
-
                 $fiscal_code = $_POST['cf'];
                 $last_name = $_POST['cognome'];
                 $first_name = $_POST['nome'];
@@ -19,36 +14,43 @@
                 $date = $_POST['data'] . ' ' . $_POST['ora'] . ':00';
                 $number_of_people = $_POST['n_persone'];
                 $notes = $_POST['note_aggiuntive'];
-                $status = $_POST['status'];
-                $tables = json_decode($_POST['tavoli_assegnati']);
-
+                
                 if (isset($_POST['id'])) {
+                    if (no_permission('mostra_prenotazioni')) {
+                        header('Location: errors/forbidden.php');
+                        exit();
+                    }
+    
                     $reservation_id = $_POST['id'];
+                    $tables = json_decode($_POST['tavoli_assegnati']);
+                    $status = $_POST['cod_status'];
 
                     edit_customer($fiscal_code, $last_name, $first_name, $phone_number, $address);
                     edit_reservation($reservation_id, $fiscal_code, $date, $number_of_people, $notes, $status);
                     delete_all_booked_tables($reservation_id);
                     book_tables($reservation_id, $tables);
+                    header('Location: ../reservations_list.php');
                 } else {
                     do {
                         $reservation_id = bin2hex(random_bytes(5));
                     } while (!empty(get_reservations($reservation_id)));
 
                     create_customer($fiscal_code, $last_name, $first_name, $phone_number, $address);
-                    create_reservation($reservation_id, $fiscal_code, $date, $number_of_people, $notes, $status);
+                    create_reservation($reservation_id, $fiscal_code, $date, $number_of_people, $notes, 1);
+                    header('Location: ../index.php');
                 }
 
                 break;
 
             case 'tables':
-                if (!isset($_SESSION['account']) || !has_permission('admin', $_SESSION['account']['cod_gruppo'])) {
+                if (no_permission('admin')) {
                     header('Location: errors/forbidden.php');
                     exit();
                 }
 
                 $table_number = $_POST['id'];
                 $number_of_seats = $_POST['n_posti'];
-                $room = $_POST['sala'];
+                $room = $_POST['cod_sala'];
 
                 if (empty(get_tables($table_number))) {
                     create_table($table_number, $number_of_seats, $room);
@@ -56,10 +58,11 @@
                     edit_table($table_number, $number_of_seats, $room);
                 }
                 
+                header('Location: ../tables_list.php');
                 break;
 
             case 'rooms':
-                if (!isset($_SESSION['account']) || !has_permission('admin', $_SESSION['account']['cod_gruppo'])) {
+                if (no_permission('admin')) {
                     header('Location: errors/forbidden.php');
                     exit();
                 }
@@ -70,22 +73,22 @@
                 if (isset($_POST['id'])) {
                     $room_id = $_POST['id'];
                     edit_dining_room($room_id, $room_name, $room_type);
-                    
                 } else {
                     create_dining_room($room_name, $room_type);
                 }
-                
+
+                header('Location: ../rooms_list.php');
                 break;
                 
             case 'accounts':
-                if (!isset($_SESSION['account']) || !has_permission('admin', $_SESSION['account']['cod_gruppo'])) {
+                if (no_permission('admin')) {
                     header('Location: errors/forbidden.php');
                     exit();
                 }
                 
                 $username = $_POST['username'];
                 $password = $_POST['password'];
-                $group = $_POST['gruppo'];
+                $group = $_POST['cod_gruppo'];
 
                 if (isset($_POST['id'])) {
                     $token = $_POST['id'];
@@ -95,6 +98,7 @@
                     create_account($username, $password, $group);
                 }
                 
+                header('Location: ../accounts_list.php');
                 break;
         }
     }
